Add a bunch of scalac flags.
https://tpolecat.github.io/2014/04/11/scalac-flags.html

+++ src/main/scala/util/BitmapTransformations.scala
@@ -9,7 +9,7 @@ object BitmapTransformations {
     val paint = new Paint()
     paint.setAntiAlias(true)
     paint.setShader(new BitmapShader(b, Shader.TileMode.CLAMP, Shader.TileMode.CLAMP))
-    canvas.drawRoundRect(new RectF(0.0f, 0.0f, b.getWidth, b.getHeight), r, r, paint)
+    canvas.drawRoundRect(new RectF(0.0f, 0.0f, b.getWidth.toFloat, b.getHeight.toFloat), r, r, paint)
     bitmapRounded
   }
 }
